* Small change to the "Registration Free COM" code. I can now access the COM based plugin's on Windows Vista and above without administrative elevation.
However manifest files generated for Windows 7 are incompatiable with
those needed for Windows XP. The issue appears to be related to the way
Activation Context API and LoadTypeLib() behave on the two different
versions of Windows, the format of the relative file path to the plugin
varies. It is not a major issue.

I'm leaving the code disabled for the moment.

Signed-off-by: swordmaster2k <b00056835@student.itb.ie>